% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempDisaggDGP.R
\name{TempDisaggDGP}
\alias{TempDisaggDGP}
\title{High and low-frequency data generating processes}
\usage{
TempDisaggDGP(
  n_l,
  m,
  p = 1,
  beta = 0.5,
  sparsity = 1,
  method = "Denton-Cholette",
  annualMat = "sum",
  rho = 0,
  mean_X = 0.5,
  sd_X = 1,
  sd_e = 1,
  simul = FALSE
)
}
\arguments{
\item{n_l}{size of the low frequency series}

\item{m}{the integer multiple for generating the high-frequency series}

\item{beta}{the positive and negative beta elements for the coefficient vector}

\item{sparsity}{sparsity percentage of the coefficient vector}

\item{method}{nominates the choice of the DGP}

\item{annualMat}{choice of the disaggregation matix}

\item{rho}{the autocorrelation coefficient}

\item{mean_X}{mean of the design matrix}

\item{sd_X}{standard deviation of the design matrix}

\item{sd_e}{standard deviation of the errors}

\item{simul}{when TRUE the design matrix and the coefficient vector are fixed}
}
\description{
This function generates the high-frequency \eqn{mn_l \times 1} response vector \eqn{y}, according to \eqn{y=X\beta+\epsilon}, where \eqn{X} is an \eqn{mn_l\times p} matrix of indicator
series, and the \eqn{mn_l\times 1} coefficient vector may be sparse. The low-frequency \eqn{n_l\times 1} vector
can be generated by pre-multiplying a disaggregation matrix \eqn{n_l\times mn_l} matrix, such that the sum, the average, the last or the first value of \eqn{y} equates the
corresponding \eqn{Y} observation (see \insertCite{sax2013temporal;textual}{HDTempDisagg}).
}
\examples{
TempDisaggDGP(n_l = 10, m = 4, p = 4, method = 'Chow-Lin', annualMat = 'sum', mean_X = 0.5, sd_X = 1, sd_e = 1 , rho = 0.5)
}
\references{
\insertAllCited{}
}
\keyword{DGP}
\keyword{high-frequency}
\keyword{low-frequency}
\keyword{sparse}
